refactor(gateway): extract transport session task
